requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.11.1
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
